+++ tests/test_io.py
@@ -168,14 +168,19 @@ class TestExtractors(TestCase):
 
             num_frames = extractor.get_num_frames()
             num_rois = extractor.get_num_rois()
+            num_background_components = extractor.get_num_background_components()
             for trace_name, trace in extractor.get_traces_dict().items():
                 if trace is None:
                     continue
                 assert trace.shape[0] == num_frames
-                assert trace.shape[1] == num_rois
-
                 assert extractor.get_traces(name=trace_name).shape[0] == num_frames
-                assert extractor.get_traces(name=trace_name).shape[1] == num_rois
+
+                if trace_name == "neuropil":
+                    assert trace.shape[1] == num_background_components
+                    assert extractor.get_traces(name=trace_name).shape[1] == num_background_components
+                else:
+                    assert trace.shape[1] == num_rois
+                    assert extractor.get_traces(name=trace_name).shape[1] == num_rois
 
         except NotImplementedError:
             return